Added a multithreading module
Added a module that enables an easy portable use of lockless
multithreaded algorithms in multicore systems.

I used a C module becase C++ 11 is notorius for synchronizations bugs in
it's thread implementation, It is also more simple and portable
in that way.

With the right design this module could increase the performance by
a factor of the number of cores in the CPU or more.

Signed-off-by: barakav <barakavigdory@gmail.com>