nj.cpp added a better update nodes function
Using Perf and Strace on rate4site I detected a severe performance
bottleneck (44% of thev times in some runs) in UpdateDistanceTableAndCurrentNodes function.
So I had to improve the performance and readability before
perpearing it for multithreaded implementation.

the new implementation is using an in-place algorithm which is far
superior to the previous one.

Signed-off-by: barakav <barakavigdory@gmail.com>